# Optimized std.datetime.timezone.SimpleTimeZone.fromISOExtString