@JackStouffer 2018-01-29
@wilzbach With counting until `T` in the string `"2012-01-01T10:10:10"` with 500 million iterations

```
$ ldc2 -O -release test.d && ./test
indexOf				2 secs, 442 ms, 141 μs, and 6 hnsecs
byCodeUnit.countUntil		2 secs, 606 ms, 702 μs, and 4 hnsecs
representation.countUntil	2 secs, 650 ms, 292 μs, and 3 hnsecs

$ dmd -O -inline -release test.d && ./test
indexOf				4 secs, 85 ms, 213 μs, and 8 hnsecs
byCodeUnit.countUntil		7 secs, 807 ms, 216 μs, and 9 hnsecs
representation.countUntil	9 secs, 220 ms, 355 μs, and 4 hnsecs
```